refactor: /infra/auth 패키지 제거
