+++ src/main/java/gg/agit/konect/infra/oauth/AppleOAuthProperties.java
@@ -1,4 +1,4 @@
-package gg.agit.konect.infra.auth.oauth;
+package gg.agit.konect.infra.oauth;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
