+++ src/main/java/gg/agit/konect/infra/oauth/AppleOAuthServiceImpl.java
@@ -1,4 +1,4 @@
-package gg.agit.konect.infra.auth.oauth;
+package gg.agit.konect.infra.oauth;
 
 import java.util.Optional;
