+++ src/main/java/gg/agit/konect/infra/oauth/KakaoOAuthServiceImpl.java
@@ -1,4 +1,4 @@
-package gg.agit.konect.infra.auth.oauth;
+package gg.agit.konect.infra.oauth;
 
 import java.util.Map;
 import java.util.Optional;
@@ -18,27 +18,26 @@ import gg.agit.konect.domain.user.repository.UserRepository;
 import gg.agit.konect.global.auth.oauth.SocialOAuthService;
 import lombok.RequiredArgsConstructor;
 
-@Service("naver")
+@Service("kakao")
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class NaverOAuthServiceImpl extends DefaultOAuth2UserService implements SocialOAuthService {
+public class KakaoOAuthServiceImpl extends DefaultOAuth2UserService implements SocialOAuthService {
 
     private final UserRepository userRepository;
     private final UnRegisteredUserRepository unRegisteredUserRepository;
 
-    @Transactional
     @Override
+    @Transactional
     public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
         OAuth2User oAuth2User = super.loadUser(userRequest);
 
-        Map<String, Object> response = oAuth2User.getAttribute("response");
-        String email = (String)response.get("email");
+        Map<String, Object> kakaoAccount = oAuth2User.getAttribute("kakao_account");
+        String email = (String)kakaoAccount.get("email");
 
         String registrationId = userRequest.getClientRegistration().getRegistrationId().toUpperCase();
         Provider provider = Provider.valueOf(registrationId);
 
         Optional<User> registered = userRepository.findByEmailAndProvider(email, provider);
-
         if (registered.isPresent()) {
             return oAuth2User;
         }
