+++ src/main/java/gg/agit/konect/infra/oauth/NaverOAuthServiceImpl.java
@@ -1,5 +1,6 @@
-package gg.agit.konect.infra.auth.oauth;
+package gg.agit.konect.infra.oauth;
 
+import java.util.Map;
 import java.util.Optional;
 
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
@@ -17,10 +18,10 @@ import gg.agit.konect.domain.user.repository.UserRepository;
 import gg.agit.konect.global.auth.oauth.SocialOAuthService;
 import lombok.RequiredArgsConstructor;
 
-@Service("google")
+@Service("naver")
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class GoogleOAuthServiceImpl extends DefaultOAuth2UserService implements SocialOAuthService {
+public class NaverOAuthServiceImpl extends DefaultOAuth2UserService implements SocialOAuthService {
 
     private final UserRepository userRepository;
     private final UnRegisteredUserRepository unRegisteredUserRepository;
@@ -29,7 +30,9 @@ public class GoogleOAuthServiceImpl extends DefaultOAuth2UserService implements 
     @Override
     public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
         OAuth2User oAuth2User = super.loadUser(userRequest);
-        String email = oAuth2User.getAttribute("email");
+
+        Map<String, Object> response = oAuth2User.getAttribute("response");
+        String email = (String)response.get("email");
 
         String registrationId = userRequest.getClientRegistration().getRegistrationId().toUpperCase();
         Provider provider = Provider.valueOf(registrationId);
